Refactor: 쓰지 않는 파일 및 코드 정리
