--- auto_stock/common/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import HealthCheckView
-
-urlpatterns = [
-    path('health/', HealthCheckView.as_view(), name='health-check'),
-]
