+++ auto_stock/common/utils/swagger.py
@@ -2,6 +2,8 @@ from rest_framework import permissions
 from drf_yasg.views import get_schema_view
 from drf_yasg import openapi
 
+
+## Swagger 스키마 정의 
 schema_view = get_schema_view(
     openapi.Info(
         title="Auto-Stock Trading Project",
