+++ auto_stock/indices/views.py
@@ -2,13 +2,15 @@
 import logging
 
 from django.http import JsonResponse
-from rest_framework import status
+from drf_yasg.utils import swagger_auto_schema
+
+from rest_framework import status, serializers
 from rest_framework.decorators import api_view, renderer_classes
 from rest_framework.renderers import JSONRenderer, BrowsableAPIRenderer
 from rest_framework.response import Response
-from drf_yasg.utils import swagger_auto_schema
-from rest_framework import serializers
+
 from .services import get_indices_payload, get_indices_realtime_payload
+
 
 logger = logging.getLogger(__name__)
