+++ auto_stock/kis_test/admin.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 
 from .models import RealtimeSymbol
 
+
 @admin.register(RealtimeSymbol)
 class RealtimeSymbolAdmin(admin.ModelAdmin):
     list_display = ("identifier", "code", "name", "updated_at")
