+++ auto_stock/trading/services/trading_strategy.py
@@ -1,5 +1,6 @@
 from trading.data.trading_data import TradeSignalParam
 
+
 def determine_signal(symbol: str, rsi: float, price: float):
     if rsi < 5:
         return TradeSignalParam(symbol, "BUY", f"RSI={rsi:.2f}", rsi, price)
