+++ auto_stock/trading/urls.py
@@ -1,9 +1,10 @@
 from django.urls import path
+
 from trading import views
 
+
 urlpatterns = [
-    # 테스트용
-    path('rsi/<str:symbol>/', views.rsi_trade_view, name='rsi-vaule'),
+    # this urls are included under the prefix: /api/trading/
     # 자동 주문 요청
     path('auto/', views.start_auto_trading, name='rsi-auto-trade'),
 ]
